+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CLI flags for config options: `--indent-width`, `--max-width`, `--quote-style`
+
 ## [0.2.0] - 2025-12-07
 
 ### Added